feat(Next > All): Show city and location details using hook
